Attempt to make the CI pass
Alpine is not very fond of shared objects, which tree-sitter is based on. tree-sitter build failed using a cargo-build tree-sitter-cli package, with the alpine-provided package it works but mdbook fails to use the resulting so files.

Easiest solution is probably to switch the CI base image to something non-musl based.